Build one wheel by using local version identifiers
